requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
Pillow==10.1.0
sqlparse==0.5.3
tzdata==2024.2
pytest==8.3.4
pytest-django==4.9.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-filter==23.2
django-cors-headers==4.2.0
python-decouple==3.8
